Move DicePreview to a shared location
